requirements.txt
opencv-python==4.8.0.76
numpy==1.23.5
scikit-image==0.21.0
dlib==19.24.2
imutils==0.5.4
pillow==9.4.0
pyyaml==6.0
scikit-learn==1.2.2
transformers==4.33.3
torch==2.0.1
pandas==1.5.3
